Code refactoring + Blender 4.1 compatibility
