+++ settings/__init__.py
@@ -1,0 +1,9 @@
+from . import settings_main
+
+
+def register():
+    settings_main.register()
+
+
+def unregister():
+    settings_main.unregister()
